Condense SQL schema.
In this commit, I condensed the SQL schema by removing everything except
the `CREATE TABLE` calls, including the character-encoding and storage-engine
metadata at the end of those calls. To ensure that the schema should still
work (though I didn't actually test loading this into an empty database), I
moved the `users` table to before the `tickets` table that references it, and
also added `CREATE DATABASE IF NOT EXISTS` and `USE` statements at the
beginning.